requirements.txt
flask
python-dotenv
google-cloud-vision
pdf2image
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai